公開準備: MySQLに変更、環境変数を利用、Dockerfileを作成

+++ db/migrate/20210504133413_create_executions.rb
@@ -1,7 +1,7 @@
 class CreateExecutions < ActiveRecord::Migration[6.1]
   def change
     create_table :executions do |t|
-      t.integer :account_id, null: false
+      t.bigint :account_id, null: false
       t.text :program, null: false
       t.text :input, null: false
       t.text :output, null: false
